requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
xgboost>=2.1.0
lightgbm>=4.5.0
pyyaml>=6.0.0
joblib>=1.4.0
httpx>=0.27.0
